Update do_report.rb
We would like to see the value of the date field (expression) in the report rather than the type of date (label). Also the colon seemed unnecessary. I assume this may have been a bug but wasn't sure of the original intention, so if I'm misunderstanding please let me know :)